% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequentist_functions.R
\name{covMatMaker}
\alias{covMatMaker}
\title{covMatMakter - calculates covariance matrix}
\usage{
covMatMaker(x.r, is.scaled = TRUE)
}
\arguments{
\item{x.r}{Reference panel}

\item{is.scaled}{Flag, was the data scaled beforehand, currently package only support TRUE}
}
\value{
list of covariance matrix and inverse covariance matrix
}
\description{
covMatMakter - calculates covariance matrix
}
