% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variance_functions.R
\name{createBootStrap}
\alias{createBootStrap}
\title{Create bootstrap for matrix}
\usage{
createBootStrap(x)
}
\arguments{
\item{x}{matrix of n rows}
}
\value{
returns a matrix with n rows sampled from x
}
\description{
Create bootstrap for matrix
}
