% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variance_functions.R
\name{createListCov}
\alias{createListCov}
\title{Create list of boostraped scaled covariance matrices}
\usage{
createListCov(x, B = nrow(x))
}
\arguments{
\item{x}{matrix of n rows}

\item{B}{integer, specifying the number of bootstrap samples to create}
}
\value{
returns a list of bootstrapped scaled covariance matrices
}
\description{
Create list of boostraped scaled covariance matrices
}
