% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variance_functions.R
\name{estimateVarAdd}
\alias{estimateVarAdd}
\title{estimateVarAdd - estimation of additional variance}
\usage{
estimateVarAdd(
  beta.mc,
  beta.omega,
  x.r,
  ind.beta.mc,
  ind.beta.omega,
  omega,
  n.o
)
}
\arguments{
\item{beta.mc}{The estimate of coefficient of joint regression (transformed marginal beta)}

\item{beta.omega}{Transformed `beta.mc` basically, just `beta.omega %*% omega`, where omega is the estimated inverse covariance matrix.}

\item{x.r}{The reference panel.}

\item{ind.beta.mc}{Indices of which `beta.mc` to consider, in the bayesian scenario determined by vector gamma.}

\item{ind.beta.omega}{Indices of which `beta.omega` to consider}

\item{omega}{- Inverse of estimated covaraince matrix}

\item{n.o}{- Number of observation in original study}
}
\value{
A list containing the varius variance terms, including the naive. `total` is the combined added variance from
using the reference panel
}
\description{
estimateVarAdd - estimation of additional variance
}
