% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variance_functions.R
\name{estimateVarAddGauss}
\alias{estimateVarAddGauss}
\title{estimateVarAddGauss - estimation of additional variance}
\usage{
estimateVarAddGauss(beta.mc, cov.mat, n.o, n.r, ind.vec)
}
\arguments{
\item{beta.mc}{The estimate of coefficient of joint regression (transformed marginal beta)}

\item{n.o}{- Number of observation in original study}

\item{cov.list.boot}{Bootstrap of scaled covariance matrices.}

\item{ind.beta.mc}{Indices of which `beta.mc` to consider, in the bayesian scenario determined by vector gamma.}

\item{ind.beta.omega}{Indices of which `beta.omega` to consider}

\item{omega}{- Inverse of estimated covaraince matrix}
}
\value{
A list containing the varius variance terms, including the naive. `total` is the combined added variance from
using the reference panel
}
\description{
estimateVarAddGauss - estimation of additional variance
}
