% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequentist_functions.R
\name{testCoef}
\alias{testCoef}
\title{testCoef - Two sided testing of coefficients}
\usage{
testCoef(est.beta, var.beta, method = "BH")
}
\arguments{
\item{est.beta}{Estimated coefficients to test}

\item{var.beta}{The variance of coefficients}

\item{method}{Method of multiplicity correction, adjusted according to}
}
\value{
A list containing the varius variance terms, including the naive. `total` is the combined added variance from
using the reference panel
}
\description{
testCoef - Two sided testing of coefficients
}
