% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variance_functions.R
\name{estimateVarAdd}
\alias{estimateVarAdd}
\title{estimateVarAdd - estimation of additional variance}
\usage{
estimateVarAdd(beta.mc, cov.list, cov.mat, n.o, n.r, ind.vec, diag.flag = F)
}
\arguments{
\item{beta.mc}{The estimate of coefficient of joint regression (transformed marginal beta)}

\item{cov.list}{List of covariance matrices}

\item{cov.mat}{Estimated covarinace matrix}

\item{n.o}{Number of observation in original study}

\item{n.r}{Number of observation in reference study}

\item{ind.vec}{Indices of which `beta.mc` to consider}

\item{diag.flag}{- Flag, to estimate the variance of the covarinace matrix based on the main diagonal, or full estimation.}
}
\value{
A list containing the varius variance terms, including the naive. `total` is the combined added variance from
using the reference panel
}
\description{
estimateVarAdd - estimation of additional variance
}
